Use -Werror in debug mode for most warnings
+ silence some warnings

+++ src/rtp/rtp.h
@@ -51,7 +51,7 @@ extern "C" {
 #endif
 
 #define RTP_VERSION 2
-#define RTP_PACKET_HEADER_SIZE ((sizeof(char *) * 2) + sizeof(uint32_t *) + (2 * sizeof(int)))
+#define RTP_PACKET_HEADER_SIZE ((int)((sizeof(char *) * 2) + sizeof(uint32_t *) + (2 * sizeof(int))))
 #define RTP_MAX_MTU 9000
 #define RTP_MAX_PACKET_LEN (RTP_MAX_MTU+RTP_PACKET_HEADER_SIZE)
